pyproject.toml
[tool.poetry]
name = "diary56x"
version = "0.0.1"
description = "Open-source digital diary for schools made with Django & Vue.JS"
authors = ["AlanTheKnight <51376126+AlanTheKnight@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.10.0"
Pillow = "^8.2.0"
django-cleanup = "^5.1.0"
django-debug-toolbar = "^3.2"
django-extensions = "^3.1.2"
django-filter = "^2.4.0"
django-widget-tweaks = "^1.4.8"
Markdown = "^3.3.4"
mkdocs = "^1.2"
poetry = "^1.1.6"
toml = "^0.10.2"
djangorestframework-simplejwt = "^5.0.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
black = "^21.11b1"
flake8 = "^4.0.1"
flake8-builtins = "^1.5.3"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.1"
flake8-quotes = "^3.3.1"
isort = "^5.10.1"
pylint = "^2.11.1"
pylint-django = "^2.4.4"
pytest = "^6.2.5"
pytest-django = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
